requirements.txt
fastapi
uvicorn
psycopg2
SQLAlchemy
python-dotenv
databases[postgresql]
alembic
pydantic[email]
passlib[bcrypt]
fastapi-mail
python-jose[cryptography]